// app.client/src/api/models/BillingItemListVM.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type BillingItemListVM = {
    billingItemId?: string | null;
    itemNum?: string | null;
    billingDescription?: string | null;
    notes?: string | null;
    isSalesTaxable?: boolean;
    isActiveRecord?: boolean;
    saleUnitPrice?: number;
    primarySalesGroup?: string | null;
    unitOfMeasure?: string | null;
    isVoided?: boolean;
    createdBy?: string | null;
    createdDate?: string;
    lastModifiedBy?: string | null;
    lastModifiedDate?: string | null;
};